compare between runs add conditions count to evaluation - interesting

+++ outputs/java/changed/AREA_OF_THE_CIRCLE_THAT_HAS_A_SQUARE_AND_A_CIRCLE_INSCRIBED_IN_IT.java
@@ -18,8 +18,8 @@ static float f_gold ( int a ) {
 
 //
 public static double f_filled ( double a ) {
-    int area = ( Math . PI * a * a ) / 4 ;
-    return area ;
+  double area = ( Math . PI * a * a ) / 4 ;
+  return area ;
 }
 
 public static void main(String args[]) {
