+++ outputs/java/changed/DYCK_PATH.java
@@ -24,12 +24,14 @@ public static int f_gold ( int n ) {
 
 //
 public static double f_filled ( int n ) {
-    int res = 1 ;
-    for ( int i = 0 ;  i <= n ;  i ++ ) {
-        res *= ( 2 * n - i ) ;
-        res /= ( i + 1 ) ;
-    }
-    return res / ( n + 1 ) ;
+  double res = 1 ;
+  for ( int i = 0 ;
+  i <= n ;
+  i ++ ) {
+    res *= ( 2 * n - i ) ;
+    res /= ( i + 1 ) ;
+  }
+  return res / ( n + 1 ) ;
 }
 
 public static void main(String args[]) {
