+++ outputs/java/changed/SUM_MATRIX_ELEMENT_ABSOLUTE_DIFFERENCE_ROW_COLUMN_NUMBERS_2.java
@@ -21,11 +21,11 @@ static int f_gold ( int n ) {
 
 //
 public static int f_filled ( int n ) {
-    n -- ;
-    int sum = 0 ;
-    sum += ( n * ( n + 1 ) ) / 2 ;
-    sum += ( n * ( n + 1 ) * ( 2 * n + 1 ) ) / 6 ;
-    return ( int ) sum ;
+  n -- ;
+  double sum = 0 ;
+  sum += ( n * ( n + 1 ) ) / 2 ;
+  sum += ( n * ( n + 1 ) * ( 2 * n + 1 ) ) / 6 ;
+  return ( int ) sum ;
 }
 
 public static void main(String args[]) {
